+++ superset/connectors/base.py
@@ -1,3 +1,7 @@
+import json
+
+from superset import utils
+
 
 class Datasource(object):
