+++ superset/connectors/druid/__init__.py
@@ -1,0 +1,2 @@
+from . import models  # noqa
+from . import views  # noqa
