+++ superset/utils.py
@@ -438,7 +438,7 @@ def pessimistic_connection_handling(target):
         cursor.close()
 
 
-class QueryStatus:
+class QueryStatus(object):
 
     """Enum-type class for query statuses"""
