+++ superset/views/__init__.py
@@ -1,0 +1,2 @@
+from . import base  # noqa
+from . import core  # noqa
